+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ## [Unreleased]
 
+## [Released]
+
+## [1.4.0] - 2021-02-09
+
 ### Added
 
 - Added support for a TOML config file
@@ -14,8 +18,6 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 ### Fixed
 
 - Fixed atomic values not being initialized
-
-## [Released]
 
 ## [1.3.0] - 2021-02-09
 
@@ -61,8 +63,9 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 [Semantic Versioning]: https://semver.org/
 
 <!-- Versions -->
-[Unreleased]: https://github.com/erri120/rpgmpacker/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/erri120/rpgmpacker/compare/v1.4.0...HEAD
 [Released]: https://github.com/erri120/rpgmpacker/releases/
+[1.4.0]: https://github.com/erri120/rpgmpacker/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/erri120/rpgmpacker/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/erri120/rpgmpacker/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/erri120/rpgmpacker/compare/v1.1.0...v1.2.0